Fixed uint64 overflows

+++ contracts/BondDepository.sol
@@ -22,7 +22,7 @@ contract BondDepository is IBondDepository, NoteKeeper {
     event CreateMarket(uint256 indexed id, address indexed baseToken, address indexed quoteToken, uint256 initialPrice);
     event CloseMarket(uint256 indexed id);
     event Bond(uint256 indexed id, uint256 amount, uint256 price);
-    event Tuned(uint256 indexed id, uint64 oldControlVariable, uint64 newControlVariable);
+    event Tuned(uint256 indexed id, uint256 oldControlVariable, uint256 newControlVariable);
 
     /* ======== STATE VARIABLES ======== */
 
@@ -139,10 +139,10 @@ contract BondDepository is IBondDepository, NoteKeeper {
         // markets keep track of how many quote tokens have been
         // purchased, and how much OHM has been sold
         market.purchased += _amount;
-        market.sold += uint64(payout_);
+        market.sold += payout_;
 
         // incrementing total debt raises the price of the next bond
-        market.totalDebt += uint64(payout_);
+        market.totalDebt += payout_;
 
         emit Bond(_id, _amount, price);
 
@@ -200,7 +200,7 @@ contract BondDepository is IBondDepository, NoteKeeper {
         if (adjustments[_id].active) {
             Adjustment storage adjustment = adjustments[_id];
 
-            (uint64 adjustBy, uint48 secondsSince, bool stillActive) = _controlDecay(_id);
+            (uint256 adjustBy, uint48 secondsSince, bool stillActive) = _controlDecay(_id);
             terms[_id].controlVariable -= adjustBy;
 
             if (stillActive) {
@@ -241,13 +241,13 @@ contract BondDepository is IBondDepository, NoteKeeper {
              * i.e. market has 10 days remaining. deposit interval is 1 day. capacity
              * is 10,000 OHM. max payout would be 1,000 OHM (10,000 * 1 / 10).
              */
-            markets[_id].maxPayout = uint64((capacity * meta.depositInterval) / timeRemaining);
+            markets[_id].maxPayout = (capacity * meta.depositInterval) / timeRemaining;
 
             // calculate the ideal total debt to satisfy capacity in the remaining time
             uint256 targetDebt = (capacity * meta.length) / timeRemaining;
 
             // derive a new control variable from the target debt and current supply
-            uint64 newControlVariable = uint64((price * treasury.baseSupply()) / targetDebt);
+            uint256 newControlVariable = uint256((price * treasury.baseSupply()) / targetDebt);
 
             emit Tuned(_id, terms[_id].controlVariable, newControlVariable);
 
@@ -256,7 +256,7 @@ contract BondDepository is IBondDepository, NoteKeeper {
             } else {
                 // if decrease, control variable change will be carried out over the tune interval
                 // this is because price will be lowered
-                uint64 change = terms[_id].controlVariable - newControlVariable;
+                uint256 change = terms[_id].controlVariable - newControlVariable;
                 adjustments[_id] = Adjustment(change, _time, meta.tuneInterval, true);
             }
             metadata[_id].lastTune = _time;
@@ -295,14 +295,14 @@ contract BondDepository is IBondDepository, NoteKeeper {
          *
          * 1e18 = ohm decimals (9) + initial price decimals (9)
          */
-        uint64 targetDebt = uint64(_booleans[0] ? ((_market[0] * 1e18) / _market[1]) / 10**decimals : _market[0]);
+        uint256 targetDebt = _booleans[0] ? ((_market[0] * 1e18) / _market[1]) / 10**decimals : _market[0];
 
         /*
          * max payout is the amount of capacity that should be utilized in a deposit
          * interval. for example, if capacity is 1,000 OHM, there are 10 days to conclusion,
          * and the preferred deposit interval is 1 day, max payout would be 100 OHM.
          */
-        uint64 maxPayout = uint64((targetDebt * _intervals[0]) / secondsToConclusion);
+        uint256 maxPayout = (targetDebt * _intervals[0]) / secondsToConclusion;
 
         /*
          * max debt serves as a circuit breaker for the market. let's say the quote
@@ -343,10 +343,10 @@ contract BondDepository is IBondDepository, NoteKeeper {
         terms.push(
             Terms({
                 fixedTerm: _booleans[1],
-                controlVariable: uint64(controlVariable),
+                controlVariable: controlVariable,
                 vesting: uint48(_terms[0]),
                 conclusion: uint48(_terms[1]),
-                maxDebt: uint64(maxDebt)
+                maxDebt: maxDebt
             })
         );
 
@@ -448,12 +448,12 @@ contract BondDepository is IBondDepository, NoteKeeper {
      * @param _id          ID of market
      * @return             amount of debt to decay
      */
-    function debtDecay(uint256 _id) public view override returns (uint64) {
+    function debtDecay(uint256 _id) public view override returns (uint256) {
         Metadata memory meta = metadata[_id];
 
         uint256 secondsSince = block.timestamp - meta.lastDecay;
 
-        return uint64((markets[_id].totalDebt * secondsSince) / meta.length);
+        return (markets[_id].totalDebt * secondsSince) / meta.length;
     }
 
     /**
@@ -463,7 +463,7 @@ contract BondDepository is IBondDepository, NoteKeeper {
      * @return             control variable for market in OHM decimals
      */
     function currentControlVariable(uint256 _id) public view returns (uint256) {
-        (uint64 decay, , ) = _controlDecay(_id);
+        (uint256 decay, , ) = _controlDecay(_id);
         return terms[_id].controlVariable - decay;
     }
 
@@ -553,7 +553,7 @@ contract BondDepository is IBondDepository, NoteKeeper {
         internal
         view
         returns (
-            uint64 decay_,
+            uint256 decay_,
             uint48 secondsSince_,
             bool active_
         )
